Avoid writing the lock config file if there are no changes (#187)
* When using vendir sync -l, if the lock file does not change, then do not write the lock file.

This preserves the original modified time, which makes it work much better with tools like Make

+++ pkg/vendir/config/lock_config.go
@@ -4,7 +4,10 @@
 package config
 
 import (
+	"bytes"
+	"errors"
 	"fmt"
+	"io/fs"
 	"os"
 	"path/filepath"
 
@@ -51,14 +54,21 @@ func NewLockConfigFromBytes(bs []byte) (LockConfig, error) {
 }
 
 func (c LockConfig) WriteToFile(path string) error {
+	existingBytes, err := os.ReadFile(path)
+	if err != nil && !errors.Is(err, fs.ErrNotExist) {
+		return fmt.Errorf("Failed to check existing lock file: %w", err)
+	}
+
 	bs, err := c.AsBytes()
 	if err != nil {
 		return fmt.Errorf("Marshaling lock config: %s", err)
 	}
 
-	err = os.WriteFile(path, bs, 0600)
-	if err != nil {
-		return fmt.Errorf("Writing lock config: %s", err)
+	if bytes.Compare(existingBytes, bs) != 0 {
+		err = os.WriteFile(path, bs, 0600)
+		if err != nil {
+			return fmt.Errorf("Writing lock config: %s", err)
+		}
 	}
 
 	return nil
